+++ apps/native-component-list/src/constants/index.ts
@@ -2,3 +2,4 @@ export { default as Alerts } from './Alerts';
 export { default as Colors } from './Colors';
 export { default as Layout } from './Layout';
 export { default as Symbols } from './Symbols';
+export { default as AndroidSymbols } from './AndroidSymbols';