+++ packages/expo-symbols/src/SymbolView.tsx
@@ -1,38 +1,40 @@
-import { Platform, requireNativeViewManager } from 'expo-modules-core';
-import { processColor } from 'react-native';
+import { useFonts } from '@expo-google-fonts/material-symbols';
+import { useMemo } from 'react';
+import { Platform, PlatformColor, Text } from 'react-native';
 
-import { NativeSymbolViewProps, SymbolViewProps } from './SymbolModule.types';
+import { SymbolViewProps } from './SymbolModule.types';
+import { androidSymbolToString } from './android';
+import regular from './android/weights/regular';
 
-let NativeView: React.ComponentType<NativeSymbolViewProps> | null = null;
+// trying to mirror iOS implementation
+const DEFAULT_SYMBOL_COLOR =
+  Platform.OS === 'android' ? PlatformColor('@android:color/system_primary_dark') : '#7d9bd4';
 
-if (Platform.OS === 'ios') {
-  NativeView = requireNativeViewManager('SymbolModule');
+function getFont(weight: SymbolViewProps['weight']) {
+  const platformWeight = typeof weight === 'object' ? weight.android : null;
+  if (!platformWeight) return regular;
+
+  return platformWeight;
 }
 
 export function SymbolView(props: SymbolViewProps) {
-  if (!NativeView) {
+  const font = useMemo(() => getFont(props.weight), [props.weight]);
+  useFonts({ [font.name]: font.font });
+  const name =
+    typeof props.name === 'object'
+      ? props.name[Platform.OS === 'android' ? 'android' : 'web']
+      : null;
+  if (!name) {
     return <>{props.fallback}</>;
   }
-
-  const nativeProps = getNativeProps(props);
-  return <NativeView {...nativeProps} />;
-}
-
-function getNativeProps(props: SymbolViewProps): NativeSymbolViewProps {
-  const colors = Array.isArray(props.colors) ? props.colors : props.colors ? [props.colors] : [];
-  const animated = !!props.animationSpec || false;
-  const type = props.type || 'monochrome';
-  const size = props.size || 24;
-  const style = props.style
-    ? [{ width: size, height: size }, props.style]
-    : { width: size, height: size };
-
-  return {
-    ...props,
-    style,
-    colors: colors.map((c) => processColor(c)),
-    tint: processColor(props.tintColor),
-    animated,
-    type,
-  };
+  return (
+    <Text
+      style={{
+        fontFamily: font.name,
+        color: props.tintColor ?? DEFAULT_SYMBOL_COLOR,
+        fontSize: props.size ?? 24,
+      }}>
+      {androidSymbolToString(name)}
+    </Text>
+  );
 }